+++ integration/autonomic_acceptance/config/test.exs
@@ -2,6 +2,7 @@ import Config
 
 config :logger, level: :warning
 config :autonomic, automatic_verification: false
+config :autonomic_typesafe, autostart: false
 
 config :autonomic_postgres, Autonomic.Store.Repo,
   url: