+++ packages/autonomic_typesafe/config/test.exs
@@ -2,3 +2,4 @@ import Config
 
 config :logger, level: :warning
 config :autonomic, automatic_verification: false
+config :autonomic_typesafe, autostart: false